+++ public/index.php
@@ -8,3 +8,5 @@ if (file_exists(__DIR__ . '/../config/db.php')) {
 }
 require_once __DIR__ . '/../config/config.php';
 require_once __DIR__ . '/../src/routing.php';
+
+session_start();
